Split schema into public and admin
