+++ app/controllers/graphql/public_controller.rb
@@ -1,0 +1,14 @@
+class Graphql::PublicController < GraphqlController
+  def execute
+    result = PublicSchema.execute(
+      params[:query],
+      variables: prepare_variables(params[:variables]),
+      context: {},
+      operation_name: params[:operationName]
+    )
+    render json: result
+  rescue StandardError => e
+    raise e unless Rails.env.development?
+    handle_error_in_development(e)
+  end
+end
