+++ app/graphql/admin/types/query_type.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Admin
+  module Types
+    class QueryType < ::Types::BaseObject
+    end
+  end
+end
