+++ app/graphql/admin_schema.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Admin schema
+class AdminSchema < GraphQL::Schema
+  mutation(Admin::Types::MutationType)
+  query(Admin::Types::QueryType)
+end
