+++ app/graphql/public/types/coffee_shop_type.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Public
+  module Types
+    class CoffeeShopType < ::Types::BaseObject
+      field :id, ID, null: false
+      field :name, String, null: false
+      field :nearby, Boolean
+      field :shop_details, String, null: false
+
+      # def shop_details
+      #   "#{object.location.address}, open until #{object.close_at}"
+      # end
+    end
+  end
+end
