+++ app/graphql/public/types/location_input.rb
@@ -1,0 +1,9 @@
+
+module Public
+  module Types
+    class LocationInput < ::Types::BaseInputObject
+      argument :x, Float
+      argument :y, Float
+    end
+  end
+end
