--- app/graphql/types/location_input.rb
@@ -1,6 +1,0 @@
-module Types
-  class LocationInput < Types::BaseInputObject
-    argument :x, Float
-    argument :y, Float
-  end
-end
