--- app/graphql/types/mutation_type.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  class MutationType < Types::BaseObject
-    # TODO: remove me
-    field :test_field, String, null: false,
-      description: "An example field added by the generator"
-    def test_field
-      "Hello World"
-    end
-  end
-end
