+++ config/routes.rb
@@ -1,5 +1,6 @@
 Rails.application.routes.draw do
-  post "/graphql", to: "graphql#execute"
+  post "/graphql/admin", to: "graphql/admin#execute"
+  post "/graphql/public", to: "graphql/public#execute"
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
